Refactor: Replace conditional with polymorphism in XML and created abstraction of TokenHandler
