+++ src/main/java/org/json/BangTokenHandler.java
@@ -1,0 +1,42 @@
+package org.json;
+
+class BangTokenHandler extends TokenHandler {
+    @Override
+    boolean handleToken(XMLTokener x, JSONObject context, XMLParserConfiguration config, int currentNestingDepth, Object token, String name) throws JSONException {
+        char c;
+        int i;
+        String string;
+        c = x.next();
+            if (c == '-') {
+                if (x.next() == '-') {
+                    x.skipPast("-->");
+                    return false;
+                }
+                x.back();
+            } else if (c == '[') {
+                token = x.nextToken();
+                if ("CDATA".equals(token)) {
+                    if (x.next() == '[') {
+                        string = x.nextCDATA();
+                        if (string.length() > 0) {
+                            context.accumulate(config.getcDataTagName(), string);
+                        }
+                        return false;
+                    }
+                }
+                throw x.syntaxError("Expected 'CDATA['");
+            }
+            i = 1;
+            do {
+                token = x.nextMeta();
+                if (token == null) {
+                    throw x.syntaxError("Missing '>' after '<!'.");
+                } else if (token == XML.LT) {
+                    i += 1;
+                } else if (token == XML.GT) {
+                    i -= 1;
+                }
+            } while (i > 0);
+            return false;
+    }
+}
