+++ src/main/java/org/json/XML.java
@@ -9,6 +9,8 @@ import java.io.StringReader;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Iterator;
+import java.util.Map;
+import java.util.HashMap;
 
 /**
  * This provides static methods to convert an XML text into a JSONObject, and to
@@ -244,15 +246,11 @@ public class XML {
      * @return true if the close tag is processed.
      * @throws JSONException Thrown if any parsing error occurs.
      */
-    private static boolean parse(XMLTokener x, JSONObject context, String name, XMLParserConfiguration config, int currentNestingDepth)
+    public static boolean parse(XMLTokener x, JSONObject context, String name, XMLParserConfiguration config, int currentNestingDepth)
             throws JSONException {
-        char c;
-        int i;
-        JSONObject jsonObject = null;
-        String string;
-        String tagName;
+        
+        // String string;
         Object token;
-        XMLXsiTypeConverter<?> xmlXsiTypeConverter;
 
         // Test for and skip past these forms:
         // <!-- ... -->
@@ -266,234 +264,21 @@ public class XML {
 
         token = x.nextToken();
 
-        // <!
+        Map<Object, TokenHandler> handlerMap = new HashMap<>();
+        handlerMap.put(BANG, new BangTokenHandler());
+        handlerMap.put(QUEST, new QuestTokenHandler());
+        handlerMap.put(SLASH, new SlashTokenHandler());
+        handlerMap.put(Character.class, new SlashTokenHandler());
 
-        if (token == BANG) {
-            c = x.next();
-            if (c == '-') {
-                if (x.next() == '-') {
-                    x.skipPast("-->");
-                    return false;
-                }
-                x.back();
-            } else if (c == '[') {
-                token = x.nextToken();
-                if ("CDATA".equals(token)) {
-                    if (x.next() == '[') {
-                        string = x.nextCDATA();
-                        if (string.length() > 0) {
-                            context.accumulate(config.getcDataTagName(), string);
-                        }
-                        return false;
-                    }
-                }
-                throw x.syntaxError("Expected 'CDATA['");
-            }
-            i = 1;
-            do {
-                token = x.nextMeta();
-                if (token == null) {
-                    throw x.syntaxError("Missing '>' after '<!'.");
-                } else if (token == LT) {
-                    i += 1;
-                } else if (token == GT) {
-                    i -= 1;
-                }
-            } while (i > 0);
-            return false;
-        } else if (token == QUEST) {
+        TokenHandler tokenHandler = handlerMap.get(token);
 
-            // <?
-            x.skipPast("?>");
-            return false;
-        } else if (token == SLASH) {
-
-            // Close tag </
-
-            token = x.nextToken();
-            if (name == null) {
-                throw x.syntaxError("Mismatched close tag " + token);
-            }
-            if (!token.equals(name)) {
-                throw x.syntaxError("Mismatched " + name + " and " + token);
-            }
-            if (x.nextToken() != GT) {
-                throw x.syntaxError("Misshaped close tag");
-            }
-            return true;
-
-        } else if (token instanceof Character) {
-            throw x.syntaxError("Misshaped tag");
-
-            // Open tag <
-
-        } else {
-            tagName = (String) token;
-            token = null;
-            jsonObject = new JSONObject();
-            boolean nilAttributeFound = false;
-            xmlXsiTypeConverter = null;
-            for (;;) {
-                if (token == null) {
-                    token = x.nextToken();
-                }
-                // attribute = value
-                if (token instanceof String) {
-                    string = (String) token;
-                    token = x.nextToken();
-                    if (token == EQ) {
-                        token = x.nextToken();
-                        if (!(token instanceof String)) {
-                            throw x.syntaxError("Missing value");
-                        }
-
-                        if (config.isConvertNilAttributeToNull()
-                                && NULL_ATTR.equals(string)
-                                && Boolean.parseBoolean((String) token)) {
-                            nilAttributeFound = true;
-                        } else if(config.getXsiTypeMap() != null && !config.getXsiTypeMap().isEmpty()
-                                && TYPE_ATTR.equals(string)) {
-                            xmlXsiTypeConverter = config.getXsiTypeMap().get(token);
-                        } else if (!nilAttributeFound) {
-                            jsonObject.accumulate(string,
-                                    config.isKeepStrings()
-                                            ? ((String) token)
-                                            : stringToValue((String) token));
-                        }
-                        token = null;
-                    } else {
-                        jsonObject.accumulate(string, "");
-                    }
-
-
-                } else if (token == SLASH) {
-                    // Empty tag <.../>
-                    if (x.nextToken() != GT) {
-                        throw x.syntaxError("Misshaped tag");
-                    }
-                    if (config.getForceList().contains(tagName)) {
-                        // Force the value to be an array
-                        if (nilAttributeFound) {
-                            context.append(tagName, JSONObject.NULL);
-                        } else if (jsonObject.length() > 0) {
-                            context.append(tagName, jsonObject);
-                        } else {
-                            context.put(tagName, new JSONArray());
-                        }
-                    } else {
-                        if (nilAttributeFound) {
-                            context.accumulate(tagName, JSONObject.NULL);
-                        } else if (jsonObject.length() > 0) {
-                            context.accumulate(tagName, jsonObject);
-                        } else {
-                            context.accumulate(tagName, "");
-                        }
-                    }
-                    return false;
-
-                } else if (token == GT) {
-                    // Content, between <...> and </...>
-                    for (;;) {
-                        token = x.nextContent();
-                        if (token == null) {
-                            if (tagName != null) {
-                                throw x.syntaxError("Unclosed tag " + tagName);
-                            }
-                            return false;
-                        } else if (token instanceof String) {
-                            string = (String) token;
-                            if (string.length() > 0) {
-                                if(xmlXsiTypeConverter != null) {
-                                    jsonObject.accumulate(config.getcDataTagName(),
-                                            stringToValue(string, xmlXsiTypeConverter));
-                                } else {
-                                    jsonObject.accumulate(config.getcDataTagName(),
-                                            config.isKeepStrings() ? string : stringToValue(string));
-                                }
-                            }
-
-                        } else if (token == LT) {
-                            // Nested element
-                            if (currentNestingDepth == config.getMaxNestingDepth()) {
-                                throw x.syntaxError("Maximum nesting depth of " + config.getMaxNestingDepth() + " reached");
-                            }
-
-                            if (parse(x, jsonObject, tagName, config, currentNestingDepth + 1)) {
-                                if (config.getForceList().contains(tagName)) {
-                                    // Force the value to be an array
-                                    if (jsonObject.length() == 0) {
-                                        context.put(tagName, new JSONArray());
-                                    } else if (jsonObject.length() == 1
-                                            && jsonObject.opt(config.getcDataTagName()) != null) {
-                                        context.append(tagName, jsonObject.opt(config.getcDataTagName()));
-                                    } else {
-                                        context.append(tagName, jsonObject);
-                                    }
-                                } else {
-                                    if (jsonObject.length() == 0) {
-                                        context.accumulate(tagName, "");
-                                    } else if (jsonObject.length() == 1
-                                            && jsonObject.opt(config.getcDataTagName()) != null) {
-                                        context.accumulate(tagName, jsonObject.opt(config.getcDataTagName()));
-                                    } else {
-                                        if (!config.shouldTrimWhiteSpace()) {
-                                            removeEmpty(jsonObject, config);
-                                        }
-                                        context.accumulate(tagName, jsonObject);
-                                    }
-                                }
-
-                                return false;
-                            }
-                        }
-                    }
-                } else {
-                    throw x.syntaxError("Misshaped tag");
-                }
-            }
+        if(tokenHandler == null) {
+            tokenHandler = new DefaultTokenHandler();
         }
-    }
-    /**
-     * This method removes any JSON entry which has the key set by XMLParserConfiguration.cDataTagName
-     * and contains whitespace as this is caused by whitespace between tags. See test XMLTest.testNestedWithWhitespaceTrimmingDisabled.
-     * @param jsonObject JSONObject which may require deletion
-     * @param config The XMLParserConfiguration which includes the cDataTagName
-     */
-    private static void removeEmpty(final JSONObject jsonObject, final XMLParserConfiguration config) {
-        if (jsonObject.has(config.getcDataTagName()))  {
-            final Object s = jsonObject.get(config.getcDataTagName());
-            if (s instanceof String) {
-                if (isStringAllWhiteSpace(s.toString())) {
-                    jsonObject.remove(config.getcDataTagName());
-                }
-            }
-            else if (s instanceof JSONArray) {
-                final JSONArray sArray = (JSONArray) s;
-                for (int k = sArray.length()-1; k >= 0; k--){
-                    final Object eachString = sArray.get(k);
-                    if (eachString instanceof String) {
-                        String s1 = (String) eachString;
-                        if (isStringAllWhiteSpace(s1)) {
-                            sArray.remove(k);
-                        }
-                    }
-                }
-                if (sArray.isEmpty()) {
-                    jsonObject.remove(config.getcDataTagName());
-                }
-            }
-        }
+        
+        return tokenHandler.handleToken(x, context, config, currentNestingDepth, token, name);  
     }
 
-    private static boolean isStringAllWhiteSpace(final String s) {
-        for (int k = 0; k<s.length(); k++){
-            final char eachChar = s.charAt(k);
-            if (!Character.isWhitespace(eachChar)) {
-                return false;
-            }
-        }
-        return true;
-    }
 
     /**
      * direct copy of {@link JSONObject#stringToNumber(String)} to maintain Android support.
